Refactor and improve the MokshaMiddleware.load_models method.
This method will now first looks to see if your application has a
``sqlalchemy.url`` set in it's configuration file, and will create a SQLAlchemy
engine with it.  If it does not exist, Moksha will create an engine for your
application based on the ``app_db`` configuration, which defaults to
``sqlite:///$APPNAME.db``.