Add an initial implementation of a dynamic, live feed tree application.
There are currently two implementations of this application, an `ajax`
version and a `live` version.  The ajax version makes a new request to our
WSGI server each time, where as the `live` implementation communicates over a
persistent Stomp-driven Orbited TCPSocket.   The live widget will
automatically connect up to a unique message topic that it uses to
communicate with the Moksha Feed Consumer in the Moksha Hub.