Remove the LiveWidget StompWidget injection hack (#28)
We no longer append a StompWidget to the bottom of each LiveWidget's template.
This was a bad idea for many reasons, and prevented multiple live widgets
from being rendered in the same request.

The new method requires adding the stomp widget to the `moksha.global` entry
point, and ensuring you're rendering Moksha's global resource widget.