Inject the moksha.global_resources function into the tmpl_context, not the widget itself.
Also, don't register the widget's resources in the BaseController anymore I'm
pretty sure we don't need to do this anymore, since We used to make the
assumption that people will not render the global resource injection widget
themselves, but I think it's essential now.  The reason for this being that we
need to inject not just JS/CSS, but also rendered widgets as well.  Since
ToscaWidgets.register_resources doesn't render widgets, it will need to be done
in the templates. This is now handled by the moksha.global_resources() method,
which returns a rendered widget for Genshi or Mako.  We could eventually work
around this by maybe having another layer of middleware at the top of the stack
that would inject rendered widgets.