Update our MokshaContainer API to match the functionality of the mbContainers.
This adds a lot of useful callbacks, and will allow us to properly handle live
widgets.  Right now when you close a LiveWidget, it will break JavaScript for
everything :(  This will allow us to detect when a window is closed, collapsed,
or iconified, and let the widget unsubscribe from any message topics.