An important LiveWidget improvement.
This allows for more than one Moksha Socket to be rendered at the same time on
a page without them stepping on eachother.  I hit a problem in Fedora
Community when 1 socket would be requesting the orbited.js, and another socket
is trying to use the Stomp connection.

This patch will make the socket detect if the other is busy, and it will queue
up it's callbacks for when the `moksha.socket_ready` event is triggered.