Start the AMQP local queue after every topic subscription.
Calling queue.start with kamaloka's AMQP bindings used to cause problems, but was
fixed in:
https://fedorahosted.org/kamaloka-js/changeset/d715bfad501f22441651bb40a6cf09cc172b7a31

This patch also will cause the AMQPSocket to not call queue.start if there are no
topics to subscribe to.  This is why we are calling it after every subscription,
which essentially becomes a noop.