Add an initial `moksha send <topic> <message>` command.
This has a few drawbacks.  I couldn't figure out how to get the message to
flush properly without calling reactor.run(), so we first schedule a task to
run 0.2 seconds in the future that stops the reactor. This feels dirty and
should eventually be improved.  It also spits out fatal connection upon
completion when using the STOMP messaging protocol.