Adapt the "._initialized" pattern for Producers.
The moksha consumers have a pattern where if the user's consumer does
not successfully call super(..., self).__init__(hub), then our
moksha.hub.api.Consumer base class never gets the opportunity to set an
`self._initialized` flag to True.  The moksha-hub looks for this flag
after it tries to initialize the consumers, and if it is absent it takes
note.

This is useful for consumers that want to decide if they should be
enabled or not.  In fedmsg for instance, we want some consumers only to
load if certain config variables are present.

This commit adapts that same logic to the Producers so they can take
advantage of the same process.