Track a count of exceptions and log to monitoring.
This will count the number of exceptions hit in a row and write that
count out with each poll to the monitoring socket.

This will be useful for a nagios check.  Sometimes, fedmsg consumers
will start raising an exception for every message they receive.  This is
typically due to a restart of our db host:  our fedmsg consumers are not
smart and they don't know how to reconnect smoothly.  They need to be
restarted.

By keeping this exception count stat, we'll be able to easily see when
that occurs and respond appropriately.