Configurable permissions on monitoring socket.
We have the problem in the Fedora Infrastructure production environment
where moksha needs to be able to write to the socket, but nrpe needs to
also be able to write to the socket (to set its subscription).
Currently, we have the socket owned by fedmsg, but the gid is set to
nrpe.  The default permissions, however, are for read-only access on the
group bits.

This change will allow us to tell moksha to set the permissions on the
socket to 0770 at startup, and thus will cut down on UNKNOWN responses
from nrpe.