Make polling producers advertise the last time they ran.
We can use this in nagios checks.  For instance, we have a
DigestProducer in FMN that every 10 seconds checks to see if anyone
should be sent a batch digest of some messages.  If it hasn't run in the
last 2 minutes, then we have a problem (and it would be nice if we
could get an alert).