Fix exception counting.
This hasn't worked in a while, and would be quite useful.  It is
supposed to be a feature whereby a moksha hub can log out stats about
how many exceptions it is encountering.  We can graph those in collectd
or trigger on them with nagios alerts.

The way it used to work was that:

1) a message would come in
2) it would be passed off to the ``consume`` implementation
3) if that raised an exception, we would catch it, increment our
   counter, and then re-raise it.

However, a few releases ago, we introduced this
multiple-threads-per-consumer pattern internally where there's a master
thread receiving messages, and it puts those messages on a work queue
for ``N`` worker threads that pull from there.  We had to move around a
bunch of stuff to make this work.. but we forgot to move exception
counting.

It has been incorrectly organized like this:

1) a message would come in
2) we would put it on the work queue
3) if putting it on the work queue raised an exception, we could catch
   it and increment.. but of course, we never get an exception just from
   putting some json in a work queue.

This commit moves the exception counting from the master thread to the
worker threads themselves, so they should actually count things from
the end ``.consume`` implementation.