Less STOMP logging.
All of that heartbeat-related logging was an artifact of when it was
being developed.  It is pretty noisy in a real deployment when it gets
bounced around between failover nodes and keeps re-negotiating the
heartbeat interval.  Best to tune that down to debug for when we need to
dig deeper.