Pass STOMP headers along to consumers.
I'd like to have this to try and debug fedora-infra/bugzilla2fedmsg#7.

Currently, the STOMP messages have a 'messageId' header that lets us
track them uniquely.  I want to figure out if it is *our* bridge that is
duplicating messages, or if it is the ActiveMQ brokers, or if it is
bugzilla itself.  It's hard to tell at the moment, because we lose all
of this routing information.  I'm going to embed the headers in the
fedmsg message body itself, so we can go back and debug later.