Bugfix for blocking mode.
Blocking mode was introduced in #39, which made it so that messages
would not be pulled into internal queues off the external bus, but
instead handled directly, "blocking" execution.  This is desirable when
working with durable queues.

The "consumer" API was addressed in #39, but I missed this spot where
messages are slotted into twisted deferreds that sit in memory before
being passed to the consumer api.