Fix Python 3 related ConfigParser issue.
For some reason, in Python 3, values from config file are not
replaced by default values given to constructor. It means that
for example %(here)s is not replaced by path to folder with
config file and `EnvironmentConfigParser._interpolate()` is trying
to find this name in environment variables.

This fix is based on manual look into ConfigParser defaults and
manual replacing default values in case they are not in
environment variables.