STOMP: wait at most 60 seconds between reconnect attempts
The STOMP extension uses an exponential backoff between reconnect
attempts when failing over between brokers. In a situation when no
brokers have been available for an extended period, this can lead
to an unreasonably long backoff, and cause unnecessary reconnect
delays once a broker becomes available. This change limits the
delay to 60 seconds.