prevent a possible memory leak in Consumer
The moksha.api.hub.consumer.Consumer class would record the time
taken to process every message by appending it to the _times list.
It assumed the MonitoringProducer would call __json__() every 5
seconds to read out the values and reset the list. If the
MonitoringProducer was not configued to run, which is common,
then the _times list grows forever, causing a memory leak.

This change replaces the list with a deque, with a maximum size of
1024 entries.