Apply stomp_queue later.
This fixes a bug we see in hubs that declare a `stomp_queue`.

The `stomp_queue` is supposed to override the individually declared topics of
all registered Consumers.  The bug is visible in most logs where the hub will
correctly connect to the declared `stomp_queue` but will *also* connect to the
consumers' topics.

This happens because `self._topics` was set in the `__init__` method, but the
hub calls back to the `.subscribe()` method shortly after initialization,
appending to the supposedly overwritten `self._topics` list.

The code change here sets `self._topics` strictly to `[stomp_queue]` in the
`connected()` method, which should win.