stomp: be more forgiving about non-utf-8 content
It's possible for clients to send data in non-utf-8 format. Right now,
this will cause the hub to traceback on message reception, and a
disconnect and failover. The traceback happens too early to inspect the
message content, so it is impossible to know which producer would need
to be fixed. This change allows the hub to receive non-utf-8 messages.
Messages in the database can be inspected for the Unicode replacement
character, to help fix misbehaving producers.