Request hub shutdown when we receive a STOMP error.
This should get us out of a situation we're in now where AMQ brokers sometimes
deny our subscription.  Moksha acknowledges (logs) this error, but then never
tries to resubscribe or reconnect or anything.  Killing the process is the
sensible thing to do, so either a user or systemd or openshift can decide to
restart the process or whatever is needed.