Manage stomp heartbeat in a separate thread.
Hopefully, this will resolve all kinds of issues we've been seeing with stomp
heartbeat management.  With the call to `callLater`, the same thread is used
but unfortunately the call is often made *after* the `interval` elapses, even
when the queue is not that busy.

By putting heartbeat events in a separate thread, hopefully the hub can keep up
with the promise it made to the broker.