stomp.py: use six.text_type() instead of decode()
b60f601 added support for SNI to the STOMP protocol. The optionsForClientTLS()
API states that the hostname argument must be in unicode format, but under
Python 3 the argument is a str which doesn't provide a decode() method. Use
six.text_type() to ensure the proper type is always used.

This was reported in #69.